move static types defitions to `quote!`
