+++ telamon-gen/Cargo.toml
@@ -26,12 +26,12 @@ lazy_static = "1.0.0"
 libc = "0.2.40"
 errno = "0.2.4"
 log = "0.4.1"
+pathfinding = "0.8.0"
+proc-macro2 = "0.4.9"
+quote = "0.6.4"
 regex = "0.2.10"
 rustfmt = "0.10.0"
 serde = "1.0.43"
-# cbindgen, used in telamon_capi which transitively depends on telamon_gen,
-# currently needs version 1.0.21 to work, so we use it here as well. See
-# https://github.com/eqrion/cbindgen/pull/159
 serde_derive = "=1.0.21"
 serde_json = "1.0.16"
 indexmap = { version = "1.0", features = ["serde-1"] }
