+++ telamon-gen/src/print/template/main.rs
@@ -145,578 +145,4 @@ pub fn init_domain_partial(store: &mut DomainStore,
 {{#each triggers~}}
     {{>trigger_check}}
 {{/each~}}
-
 // TODO(cleanup): generate (IrInstance, Domain) pair here.
-
-pub trait Domain: Copy + Eq {
-    /// Indicates if the domain is empty.
-    fn is_failed(&self) -> bool;
-    /// Indicates if the domain contains a single alternative.
-    fn is_constrained(&self) -> bool;
-    /// Indicates if the domain contains another.
-    fn contains(&self, other: Self) -> bool;
-    /// Restricts the domain to the intersection with `other`.
-    fn restrict(&mut self, other: Self);
-
-    /// Indicates if the domain has an alternatve in common with `other`.
-    fn intersects(&self, mut other: Self) -> bool where Self: Sized {
-        other.restrict(*self);
-        !other.is_failed()
-    }
-
-    /// Indicates if the domain is equal to another domain.
-    fn is(&self, mut other: Self) -> Trivalent where Self: Sized {
-        other.restrict(*self);
-        if other.is_failed() {
-            Trivalent::False
-        } else if other == *self {
-            Trivalent::True
-        } else {
-            Trivalent::Maybe
-        }
-    }
-}
-
-/// Abstracts integer choices by a range.
-#[derive(Copy, Clone, PartialEq, Eq, Debug, Serialize, Deserialize)]
-#[repr(C)]
-pub struct Range {
-    pub min: u32,
-    pub max: u32,
-}
-
-#[allow(dead_code)]
-impl Range {
-    pub const ALL: Range = Range { min: 0, max: std::u32::MAX };
-
-    pub const FAILED: Range = Range { min: 1, max: 0 };
-
-    /// Returns the full range.
-    pub fn all() -> Self { Self::ALL }
-
-    /// Inserts alternatives into the domain.
-    pub fn insert(&mut self, other: Range) {
-        self.min = std::cmp::min(self.min, other.min);
-        self.max = std::cmp::max(self.max, other.max);
-    }
-
-    /// Returns the difference between the min and max of two ranges.
-    fn get_diff_add(&self, other: Range) -> Range {
-        Range { min: other.min - self.min, max: self.max - other.max }
-    }
-
-    /// Restricts the `Range` by applying the result of `get_diff`.
-    fn apply_diff_add(&mut self, diff: Range) {
-        self.min += diff.min;
-        self.max -= diff.max;
-    }
-
-    /// Returns the difference between the min and max of two ranges.
-    fn get_diff_mul(&self, other: Range) -> Range {
-        Range { min: other.min / self.min, max: self.max / other.max }
-    }
-
-    /// Restricts the `Range` by applying the result of `get_diff`.
-    fn apply_diff_mul(&mut self, diff: Range) {
-        self.min *= diff.min;
-        self.max /= diff.max;
-    }
-
-    fn add_add(&mut self, diff: Range) {
-        self.min += diff.min;
-        self.max += diff.max;
-    }
-
-    fn add_mul(&mut self, diff: Range) {
-        self.min *= diff.min;
-        self.max *= diff.max;
-    }
-
-    fn sub_add(&mut self, diff: Range) {
-        self.min -= diff.min;
-        self.max -= diff.max;
-    }
-
-    fn sub_mul(&mut self, diff: Range) {
-        self.min /= diff.min;
-        self.max /= diff.max;
-    }
-
-    pub fn as_fixed(&self) -> Option<u32> {
-        if self.min == self.max { Some(self.min) } else { None }
-    }
-}
-
-impl Domain for Range {
-    fn is_failed(&self) -> bool { self.min > self.max }
-
-    fn is_constrained(&self) -> bool { self.min == self.max }
-
-    fn contains(&self, other: Range) -> bool {
-        self.min <= other.min && self.max >= other.max
-    }
-
-    fn restrict(&mut self, other: Range) {
-        self.min = std::cmp::max(self.min, other.min);
-        self.max = std::cmp::min(self.max, other.max);
-    }
-}
-
-/// Abstracts integer choices by a range, but only store `min`.
-#[derive(Copy, Clone, PartialEq, Eq, Debug, Serialize, Deserialize)]
-#[repr(C)]
-pub struct HalfRange { pub min: u32 }
-
-#[allow(dead_code)]
-impl HalfRange {
-    pub const ALL: HalfRange = HalfRange { min: 0 };
-
-    /// Returns the full `HalfRange`.
-    pub fn all() -> Self { Self::ALL }
-
-    /// Inserts alternatives into the domain.
-    pub fn insert(&mut self, other: HalfRange) {
-        self.min = std::cmp::min(self.min, other.min);
-    }
-
-    /// Returns the difference between the min and max of two ranges.
-    fn get_diff_add(&self, other: HalfRange) -> HalfRange {
-        HalfRange { min: other.min - self.min }
-    }
-
-    /// Restricts the `Range` by applying the result of `get_diff`.
-    fn apply_diff_add(&mut self, diff: HalfRange) { self.min += diff.min; }
-
-    /// Returns the difference between the min and max of two ranges.
-    fn get_diff_mul(&self, other: HalfRange) -> HalfRange {
-        HalfRange { min: other.min / self.min }
-    }
-
-    /// Restricts the `Range` by applying the result of `get_diff`.
-    fn apply_diff_mul(&mut self, diff: HalfRange) { self.min *= diff.min; }
-
-    fn add_add(&mut self, diff: HalfRange) {
-        self.min += diff.min;
-    }
-
-    fn add_mul(&mut self, diff: HalfRange) {
-        self.min *= diff.min;
-    }
-
-    fn sub_add(&mut self, diff: HalfRange) {
-        self.min -= diff.min;
-    }
-
-    fn sub_mul(&mut self, diff: HalfRange) {
-        self.min /= diff.min;
-    }
-}
-
-impl Domain for HalfRange {
-    fn is_failed(&self) -> bool { false }
-
-    fn is_constrained(&self) -> bool { false }
-
-    fn contains(&self, other: HalfRange) -> bool {
-        self.min <= other.min
-    }
-
-    fn restrict(&mut self, other: HalfRange) {
-        self.min = std::cmp::max(self.min, other.min);
-    }
-}
-
-#[derive(Copy, Clone, Serialize, Deserialize)]
-pub struct NumericSet {
-    len: usize,
-    values: [u32; NumericSet::MAX_LEN],
-}
-
-// FIXME: replace num_set with `mask: u16, all: u16`.
-
-// FIXME: functions that do not need the universe:
-// - insert
-// - list
-// - is_failed
-// - is_constrained
-// - contains
-// - restrict
-// FIXME: functions to delete
-// - new_fixed
-// - restrict_to
-// - as_num_set
-// FIXME: functions that needs the universe (and that's ok)
-// - gcd/lcm
-// - min/max
-// - as_constrained
-// - complement
-// FIXME: all NumDomain methods need both universe
-// > request domain in min, max
-// > request domain in comparison
-// > request domain in constructors
-// > request domain in final_value, rename it as_constrained
-// > fix templates that use each functions
-// Problem: it is not known for code: use () instead
-// FIXME: do range and half range need a universe ?
-
-#[allow(dead_code)]
-impl NumericSet {
-    pub const MAX_LEN: usize = 16;
-
-    pub const FAILED: Self = NumericSet { len: 0, values: [0; NumericSet::MAX_LEN] };
-
-    /// Creates a new `NumericSet` containing a single value.
-    pub fn new_fixed(val: u32) -> Self {
-        let mut values = [0; Self::MAX_LEN];
-        values[0] = val;
-        NumericSet { values, len: 1 }
-    }
-
-    /// Returns the set containing all the possibilities. Assumes the universe is sorted.
-    pub fn all(univers: &[u32]) -> Self {
-        assert!(univers.len() <= NumericSet::MAX_LEN);
-        let mut values = [0; NumericSet::MAX_LEN];
-        for (v, dst) in univers.iter().cloned().zip(&mut values) { *dst = v; }
-        NumericSet { len: univers.len(), values }
-    }
-
-    /// Inserts alternatives into the domain. Both domains should be from the same
-    /// universe.
-    pub fn insert(&mut self, other: NumericSet) {
-        let mut values = [0; NumericSet::MAX_LEN];
-        let (mut idx, mut idx_self, mut idx_other) = (0, 0, 0);
-        while idx_self < self.len && idx_other < other.len {
-            let mut v = self.values[idx_self];
-            if v <= other.values[idx_other] { idx_self += 1; }
-            if v >= other.values[idx_other] {
-                v = other.values[idx_other];
-                idx_other += 1;
-            }
-            values[idx] = v;
-            idx += 1;
-        }
-        for &item in &self.values[idx_self..self.len] {
-            values[idx] = item;
-            idx += 1;
-        }
-        for &item in &other.values[idx_other..other.len] {
-            values[idx] = item;
-            idx += 1;
-        }
-        self.values = values;
-        self.len = idx;
-    }
-
-    /// Returns the greatest common divisor of possible values.
-    pub fn gcd(&self) -> u32 {
-        self.values[..self.len].iter().cloned().fold1(num::integer::gcd).unwrap_or(1)
-    }
-
-    /// Returns the least common multiple of possible values.
-    pub fn lcm(&self) -> u32 {
-        self.values[..self.len].iter().cloned().fold1(num::integer::lcm).unwrap_or(1)
-    }
-
-    /// Returns the only value in the set, if it is fully constrained.
-    pub fn as_constrained(&self) -> Option<u32> {
-        if self.len == 1 { Some(self.values[0]) } else { None }
-    }
-
-    /// Lists the possible values the domain can take.
-    pub fn list<'a>(&'a self) -> Vec<Self> {
-        self.values[0..self.len].iter().cloned().map(Self::new_fixed).collect()
-    }
-
-    fn restrict_to(&mut self, other: &[u32]) {
-        let (mut new_lhs, mut old_lhs, mut rhs) = (0, 0, 0);
-        while old_lhs < self.len && rhs < other.len() {
-            if self.values[old_lhs] > other[rhs] {
-                rhs += 1;
-            } else if self.values[old_lhs] < other[rhs] {
-                old_lhs += 1;
-            } else {
-                self.values[new_lhs] = self.values[old_lhs];
-                old_lhs += 1;
-                new_lhs += 1;
-                rhs += 1;
-            }
-        }
-        self.len = new_lhs;
-    }
-
-    fn complement(&self, universe: &[u32]) -> Self {
-        let mut values = [0; Self::MAX_LEN];
-        let mut self_idx = 0;
-        let mut new_idx = 0;
-        for &item in universe {
-            if self_idx >= self.len || item < self.values[self_idx] {
-                values[new_idx] = item;
-                new_idx += 1;
-            } else if item > self.values[self_idx] {
-                panic!("self should be contained in universe")
-            } else {
-                self_idx += 1;
-            }
-        }
-        NumericSet { values, len: new_idx }
-    }
-}
-
-impl Domain for NumericSet {
-    fn is_failed(&self) -> bool { self.len == 0 }
-
-    fn is_constrained(&self) -> bool { self.len == 1 }
-
-    fn contains(&self, other: NumericSet) -> bool {
-        let (mut lhs, mut rhs) = (0, 0);
-        while lhs < self.len && rhs < other.len {
-            if self.values[lhs] > other.values[rhs] { return false; }
-            if self.values[lhs] == other.values[rhs] { rhs += 1; }
-            lhs += 1;
-        }
-        rhs == other.len
-    }
-
-    fn restrict(&mut self, other: NumericSet) {
-        self.restrict_to(&other.values[..other.len])
-    }
-}
-
-impl PartialEq for NumericSet {
-    fn eq(&self, other: &NumericSet) -> bool {
-        if self.len != other.len { return false; }
-        for i in 0..self.len {
-            if self.values[i] != other.values[i] { return false; }
-        }
-        true
-    }
-}
-
-impl std::fmt::Debug for NumericSet {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(f, "{:?}", &self.values[..self.len])
-    }
-}
-
-impl Eq for NumericSet { }
-
-/// A domain containing integers.
-pub trait NumDomain {
-    type Universe: ?Sized;
-
-    /// Returns the maximum value in the domain.
-    fn min(&self) -> u32;
-    /// Returns the minimum value in the domain.
-    fn max(&self) -> u32;
-    /// Returns the domain as a `NumericSet`, if applicable.
-    fn as_num_set(&self) -> Option<NumericSet> { None }
-
-    /// Returns the value of the domain, if it is constrained.
-    fn final_value(&self) -> u32 {
-        assert_eq!(self.min(), self.max());
-        self.min()
-    }
-
-
-    fn lt<D: NumDomain>(&self, other: D) -> bool { self.max() < other.min() }
-
-    fn gt<D: NumDomain>(&self, other: D) -> bool { self.min() > other.max() }
-
-    fn leq<D: NumDomain>(&self, other: D) -> bool { self.max() <= other.min() }
-
-    fn geq<D: NumDomain>(&self, other: D) -> bool { self.min() >= other.max() }
-
-    fn eq<D: NumDomain>(&self, other: D) -> bool {
-        self.min() == other.max() && self.max() == other.min()
-    }
-
-    fn neq<D: NumDomain>(&self, other: D) -> bool {
-        self.min() > other.max() || self.max() < other.min()
-    }
-}
-
-/// A choice that contains integers.
-pub trait NumChoice: NumDomain {
-    /// Returns the domain containing the values of the universe greater than min.
-    fn new_gt<D: NumDomain>(universe: &Self::Universe, min: D) -> Self;
-    /// Returns the domain containing the values of the universe smaller than max.
-    fn new_lt<D: NumDomain>(universe: &Self::Universe, max: D) -> Self;
-    /// Retruns the domain containing the values of the universe greater or equal to min.
-    fn new_geq<D: NumDomain>(universe: &Self::Universe, min: D) -> Self;
-    /// Returns the domain containing the values of the universe smaller or equal to min.
-    fn new_leq<D: NumDomain>(universe: &Self::Universe, min: D) -> Self;
-    /// Returns the domain containing the values of `eq` that are also in the universe.
-    fn new_eq<D: NumDomain>(universe: &Self::Universe, eq: D) -> Self;
-}
-
-impl NumDomain for Range {
-    type Universe = ();
-
-    fn min(&self) -> u32 { self.min }
-
-    fn max(&self) -> u32 { self.max }
-}
-
-impl NumChoice for Range {
-    fn new_gt<D: NumDomain>(universe: &(), min: D) -> Self {
-        let min = min.min().saturating_add(1);
-        Range { min, .. Range::ALL }
-    }
-
-    fn new_lt<D: NumDomain>(universe: &(), max: D) -> Self {
-        let max = max.max().saturating_sub(1);
-        Range { max, .. Range::ALL }
-    }
-
-    fn new_geq<D: NumDomain>(universe: &(), min: D) -> Self {
-        Range { min: min.min(), .. Range::ALL }
-    }
-
-    fn new_leq<D: NumDomain>(universe: &(), max: D) -> Self {
-        Range { max: max.max(), .. Range::ALL }
-    }
-
-    fn new_eq<D: NumDomain>(universe: &(), eq: D) -> Self {
-        Range {
-            max: eq.max(),
-            min: eq.min(),
-        }
-    }
-}
-
-impl NumDomain for HalfRange {
-    type Universe = ();
-
-    fn min(&self) -> u32 { self.min }
-
-    fn max(&self) -> u32 { std::u32::MAX }
-}
-
-impl NumChoice for HalfRange {
-    fn new_gt<D: NumDomain>(universe: &(), min: D) -> Self {
-        let min = min.min().saturating_add(1);
-        HalfRange { min }
-    }
-
-    fn new_lt<D: NumDomain>(universe: &(), _: D) -> Self { HalfRange::ALL }
-
-    fn new_geq<D: NumDomain>(universe: &(), min: D) -> Self {
-        HalfRange { min: min.min() }
-    }
-
-    fn new_leq<D: NumDomain>(universe: &(), _: D) -> Self { HalfRange::ALL }
-
-    fn new_eq<D: NumDomain>(universe: &(), eq: D) -> Self {
-        HalfRange { min: eq.min() }
-    }
-}
-
-impl NumDomain for NumericSet {
-    type Universe = [u32];
-
-    fn min(&self) -> u32 {
-        if self.len == 0 { 1 } else { self.values[0] }
-    }
-
-    fn max(&self) -> u32 {
-        if self.len == 0 { 0 } else { self.values[self.len-1] }
-    }
-
-    fn as_num_set(&self) -> Option<NumericSet> { Some(*self) }
-
-    fn neq<D: NumDomain>(&self, other: D) -> bool {
-        if let Some(other) = other.as_num_set() {
-            let (mut self_idx, mut other_idx) = (0, 0);
-            while self_idx < self.len && other_idx < other.len {
-                if self.values[self_idx] < other.values[other_idx] {
-                    self_idx += 1;
-                } else if self.values[self_idx] > other.values[other_idx] {
-                    other_idx += 1
-                } else {
-                    return false;
-                }
-            }
-            true
-        } else {
-            self.min() > other.max() || self.max() < other.min()
-        }
-    }
-}
-
-impl NumChoice for NumericSet {
-    fn new_gt<D: NumDomain>(universe: &[u32], min: D) -> Self {
-        let mut values = [0; NumericSet::MAX_LEN];
-        let min = std::cmp::min(std::u32::MAX, min.min());
-        let start = universe.binary_search(&min).map(|x| x+1).unwrap_or_else(|x| x);
-        let len = universe.len() - start;
-        for i in 0..len { values[i] = universe[start+i]; }
-        NumericSet { values, len }
-    }
-
-    fn new_lt<D: NumDomain>(universe: &[u32], max: D) -> Self {
-        let mut values = [0; NumericSet::MAX_LEN];
-        let max = std::cmp::min(std::u32::MAX, max.max());
-        let len = universe.binary_search(&max).unwrap_or_else(|x| x);
-        for i in 0..len { values[i] = universe[i]; }
-        NumericSet { values, len }
-    }
-
-    fn new_geq<D: NumDomain>(universe: &[u32], min: D) -> Self {
-        let mut values = [0; NumericSet::MAX_LEN];
-        let min = std::cmp::min(std::u32::MAX, min.min());
-        let start = universe.binary_search(&min).unwrap_or_else(|x| x);
-        let len = universe.len() - start;
-        for i in 0..len { values[i] = universe[start+i]; }
-        NumericSet { values, len }
-    }
-
-    fn new_leq<D: NumDomain>(universe: &[u32], max: D) -> Self {
-        let mut values = [0; NumericSet::MAX_LEN];
-        let max = std::cmp::min(std::u32::MAX, max.max());
-        let len = universe.binary_search(&max).map(|x| x+1).unwrap_or_else(|x| x);
-        for i in 0..len { values[i] = universe[i]; }
-        NumericSet { values, len }
-    }
-
-    fn new_eq<D: NumDomain>(universe: &[u32], eq: D) -> Self {
-        if let Some(mut eq) = eq.as_num_set() {
-            eq.restrict_to(universe);
-            eq
-        } else {
-            let mut values = [0; NumericSet::MAX_LEN];
-            let min = std::cmp::min(std::u32::MAX, eq.min());
-            let max = std::cmp::min(std::u32::MAX, eq.max());
-            let start = universe.binary_search(&min).unwrap_or_else(|x| x);
-            let len = universe.binary_search(&max).unwrap_or_else(|x| x) - start;
-            for i in 0..len { values[i] = universe[start+i]; }
-            NumericSet { values, len }
-        }
-    }
-}
-
-impl NumDomain for u32 {
-    type Universe = ();
-
-    fn min(&self) -> u32 { *self }
-
-    fn max(&self) -> u32 { *self }
-}
-
-impl<'a> NumDomain for &'a [u32] {
-    type Universe = ();
-
-    fn min(&self) -> u32 {
-        if self.is_empty() { 1 } else { self[0] }
-    }
-
-    fn max(&self) -> u32 {
-        if self.is_empty() { 0 } else { self[self.len()-1] }
-    }
-
-    fn as_num_set(&self) -> Option<NumericSet> {
-        assert!(self.len() < NumericSet::MAX_LEN);
-        let mut values = [0; NumericSet::MAX_LEN];
-        for i in 0..self.len() { values[i] = self[i] }
-        Some(NumericSet { len: self.len(), values })
-    }
-}
